news_room_v2/requirements.txt
absl-py==2.0.0
addict==2.4.0
alembic==1.18.3
altair==6.0.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astunparse==1.6.3
atomicwrites==1.4.0
backoff==2.2.1
backports.weakref==1.0.post1
bcrypt==5.0.0
blinker==1.9.0
brotlipy==0.7.0
build==1.4.0
cachetools==5.5.2
certifi==2026.1.4
cfgv==3.4.0
chromadb==1.4.1
cloudpickle==3.1.2
clyent==1.2.2
cmdstanpy==1.2.2
colorlog==6.10.1
conda-repo-cli==1.0.75
conda-verify==3.4.2
constantly==15.1.0
cssselect==1.1.0
curl_cffi==0.13.0
datashape==0.5.4
distlib==0.3.9
distro==1.9.0
durationpy==0.10
et-xmlfile==1.1.0
filelock==3.16.1
finance-datareader==0.9.90
flatbuffers==23.5.26
fonttools==4.25.0
frozendict==2.4.7
gast==0.5.4
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.23.3
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.0.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
holidays==0.49
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
hydra-core==1.3.2
identify==2.6.3
imbalanced-learn==0.14.1
importlib_resources==6.4.0
inflection==0.5.1
joblib==1.5.3
jsonpointer==2.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kagglehub==0.4.2
kagglesdk==0.1.15
keras==2.14.0
kubernetes==35.0.0
libclang==16.0.6
lightgbm==4.6.0
llvmlite==0.46.0
Mako==1.3.10
ml-dtypes==0.2.0
mmcv==2.2.0
mmengine==0.10.7
mmh3==5.2.0
mmsegmentation==1.2.2
mtcnn==0.1.1
multitasking==0.0.12
munkres==1.1.4
narwhals==2.15.0
nodeenv==1.9.1
numba==0.63.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.24.1
opencv-python==4.8.1.78
openpyxl==3.0.10
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opt-einsum==3.3.0
optuna==4.7.0
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pandas==2.3.3
patsy==0.5.3
peewee==3.19.0
pep8==1.7.1
pi-heaan==0.5.0
playwright==1.58.0
ply==3.11
posthog==5.4.0
pre_commit==4.0.1
prettytable==3.17.0
prophet==1.1.5
proto-plus==1.27.0
protobuf==5.29.5
pyarrow==11.0.0
pyasn1-modules==0.2.8
pybase64==1.4.3
pydeck==0.9.1
PyDispatcher==2.0.5
pyee==13.0.0
pyls-spyder==0.4.0
pyparsing==3.3.2
PyPika==0.51.1
pyproject_hooks==1.2.0
PyQt5-sip==12.11.0
pyrootutils==1.0.4
python-dateutil==2.9.0.post0
python-lsp-jsonrpc==1.0.0
referencing==0.37.0
requests-oauthlib==1.3.1
rich==14.3.1
rpds-py==0.30.0
rsa==4.9
scikit-learn==1.8.0
scipy==1.13.1
shap==0.44.1
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.5
slicer==0.0.7
smmap==5.0.2
stanio==0.5.0
streamlit==1.53.1
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
termcolor==2.3.0
threadpoolctl==3.6.0
timm==1.0.24
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.2
typer==0.21.1
typing_extensions==4.15.0
uritemplate==4.2.0
uvicorn==0.40.0
uvloop==0.22.1
virtualenv==20.28.0
watchfiles==1.1.1
webencodings==0.5.1
websockets==16.0
xgboost==3.1.3
yfinance==1.0
youtube-transcript-api==1.2.4
